% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaMetrics.R
\name{EvaMetrics}
\alias{EvaMetrics}
\title{Forecasting evaluation metrics
Author: Cong Feng 
This function calculates deterministic forecasting evaluation metrics}
\usage{
EvaMetrics(anly)
}
\arguments{
\item{anly}{input data in the [forecasts, actual] format}
}
\value{
six error metrics: mape, mae, nmae, mse, rmse, and nrmse
}
\description{
Forecasting evaluation metrics
Author: Cong Feng 
This function calculates deterministic forecasting evaluation metrics
}
