% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qlearning_DMS.R
\name{Qlearning_DMS}
\alias{Qlearning_DMS}
\title{Dynamic model selection using Q-learning for load forecasting
Author: Cong Feng
Reference: 1. Feng, C. and Zhang, J., 2019, February. Reinforcement learning based
              dynamic model selection for short-term load forecasting. In 2019 IEEE
              Power & Energy Society Innovative Smart Grid Technologies Conference
              (ISGT) (pp. 1-5). IEEE.
           2. Feng, C., Sun, M. and Zhang, J., 2019. Reinforced Deterministic and
              Probabilistic Load Forecasting via $ Q $-Learning Dynamic Model Selection.
              IEEE Transactions on Smart Grid, 11(2), pp.1377-1386.
This function performs dynamic model selection using Q-learning for load forecasting}
\usage{
Qlearning_DMS(
  QMS_no_models,
  QMS_no_episodes,
  QMS_no_hour,
  QMS_init_md,
  QMS_reward_selection,
  QMS_alpha,
  QMS_gamma,
  df_learn,
  df_select
)
}
\arguments{
\item{QMS_no_models}{state and action space dimension}

\item{QMS_no_episodes}{number of episodes}

\item{QMS_no_hour}{update frequency}

\item{QMS_init_md}{starting state}

\item{QMS_reward_selection}{reward strategy}

\item{QMS_alpha}{learning rate}

\item{QMS_gamma}{discount factor}

\item{df_learn}{training data frame}

\item{df_select}{selecting data frame}
}
\value{
the Q-learning forecasting results
}
\description{
Dynamic model selection using Q-learning for load forecasting
Author: Cong Feng
Reference: 1. Feng, C. and Zhang, J., 2019, February. Reinforcement learning based
              dynamic model selection for short-term load forecasting. In 2019 IEEE
              Power & Energy Society Innovative Smart Grid Technologies Conference
              (ISGT) (pp. 1-5). IEEE.
           2. Feng, C., Sun, M. and Zhang, J., 2019. Reinforced Deterministic and
              Probabilistic Load Forecasting via $ Q $-Learning Dynamic Model Selection.
              IEEE Transactions on Smart Grid, 11(2), pp.1377-1386.
This function performs dynamic model selection using Q-learning for load forecasting
}
\references{
Feng, C. and Zhang, J., 2019, February. Reinforcement learning based
              dynamic model selection for short-term load forecasting. In 2019 IEEE
              Power & Energy Society Innovative Smart Grid Technologies Conference
              (ISGT) (pp. 1-5). IEEE.
}
