% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reward_function.R
\name{Reward_function}
\alias{Reward_function}
\title{Q-learning reward functions that are defined specifically for deterministic forecasting
Author: Cong Feng 
This function calculates rewards for Q-learning agents}
\usage{
Reward_function(data_input, rwd_strategy)
}
\arguments{
\item{data_input}{input data that contains states and actions in time series format}

\item{rwd_strategy}{reward strategy index}
}
\value{
rewards
}
\description{
Q-learning reward functions that are defined specifically for deterministic forecasting
Author: Cong Feng 
This function calculates rewards for Q-learning agents
}
